added compilation flags

+++ srcs/utils.hpp
@@ -3,12 +3,13 @@
 
 #include <string>
 #include "Settings.h"
+#include <stdio.h>
 
-std::string		readFileContent(const std::string& path);
-std::string		getFileType(const std::string& filePath);
-int				getResponseCode(const std::string& filePath);
-int				getContentLen(const std::string& header);
-const Location*	findLocationByPath(const Settings& set, const std::string& path);
+std::string		    readFileContent(const std::string& path);
+std::string		    getFileType(const std::string& filePath);
+int				    getResponseCode(const std::string& filePath);
+size_t			    getContentLen(const std::string& header);
+const Location*	    findLocationByPath(const Settings& set, const std::string& path);
 const std::string	repertoryListing(const std::string directoryPath);
 
 #endif
